src/main/java/com/example/filemanager/service/StorageService.java
package com.example.filemanager.service;

import com.example.filemanager.entity.ImageData;
import com.example.filemanager.repository.StorageRepository;
import com.example.filemanager.util.ImageUtils;
import org.springframework.stereotype.Service;
import org.springframework.web.multipart.MultipartFile;

import java.io.IOException;
import java.util.Optional;

@Service
public class StorageService {

    private final StorageRepository storageRepository;

    public StorageService(StorageRepository storageRepository) {
        this.storageRepository = storageRepository;
    }


    public String uploadImage(MultipartFile file) throws IOException {
        ImageData imageData = storageRepository.save(ImageData.builder()
                .name(file.getOriginalFilename())
                .type(file.getContentType())
                .imageData(ImageUtils.compressImage(file.getBytes())).build());

        if (imageData != null) {
            return file.getOriginalFilename();
        }
        return null;

    }

    public byte[] downloadImage(String fileName) {
        Optional<ImageData> dbImageData = storageRepository.findByName(fileName);
        return ImageUtils.deCompressImage(dbImageData.get().getImageData());
    }
}
